+++ src/lib.rs
@@ -92,13 +92,10 @@ mod constraints;
 pub mod debug;
 mod error;
 #[cfg(feature = "ffi")]
-mod ffi;
+pub mod ffi;
 mod layout;
 mod position;
 mod size;
-
-#[cfg(feature = "ffi")]
-pub use ffi::*;
 
 pub use constraints::*;
 pub use error::LayoutError;
@@ -137,18 +134,20 @@ impl std::fmt::Display for GlobalId {
 
 /// Describes how a [`Layout`] should align its children.
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq, PartialOrd, Ord)]
+#[repr(u8)]
 pub enum AxisAlignment {
     /// Place content at the start.
     #[default]
-    Start,
+    Start = 0,
     /// Place content in the center.
-    Center,
+    Center = 1,
     /// Place content at the end.
-    End,
+    End = 2,
 }
 
 /// The space between the edges of a [`Layout`] node and its content.
 #[derive(Clone, Copy, Default, PartialEq, PartialOrd, Debug)]
+#[repr(C)]
 pub struct Padding {
     /// The left padding.
     pub left: f32,
@@ -178,7 +177,7 @@ impl Padding {
         }
     }
 
-    /// Create padding with symmetric vertical and horizontal sides.
+    /// Creates a [`Padding`] with symmetric vertical and horizontal sides.
     ///
     /// # Example
     ///
@@ -192,11 +191,14 @@ impl Padding {
     /// assert_eq!(padding.left,padding.right);
     /// assert_eq!(padding.bottom,padding.top);
     /// ```
+    ///
+    /// # Panics
+    /// Panics if either of the values is negative.
     pub const fn symmetric(vertical: f32, horizontal: f32) -> Self {
         Self::new(horizontal, horizontal, vertical, vertical)
     }
 
-    /// Create a [`Padding`] with equal sides.
+    /// Creates a [`Padding`] with equal sides.
     ///
     /// # Example
     /// ```
@@ -208,11 +210,14 @@ impl Padding {
     /// assert_eq!(padding.left,padding.right);
     /// assert_eq!(padding.bottom,padding.top);
     /// ```
+    ///
+    /// # Panics
+    /// Panics if `padding` is negative.
     pub const fn all(padding: f32) -> Self {
         Self::new(padding, padding, padding, padding)
     }
 
-    /// The sum of the top and bottom padding.
+    /// Returns the sum of the top and bottom padding.
     ///
     /// # Example
     ///
@@ -227,7 +232,7 @@ impl Padding {
         self.bottom + self.top
     }
 
-    /// The sum of the left and right padding.
+    /// Returns the sum of the left and right padding.
     ///
     /// # Example
     ///
@@ -242,7 +247,7 @@ impl Padding {
         self.left + self.right
     }
 
-    /// The sum of all the padding sides.
+    /// Returns the sum of all the padding sides.
     ///
     /// # Example
     /// ```